refactor: extract offer methods
